pyproject.toml
[tool.poetry]
name = "dona-francisca-poc"
version = "0.1.0"
description = ""
authors = ["Bruno Camargo <brunocam1195@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
streamlit = "^1.43.1"
python-multipart = "^0.0.20"
langchain = "^0.3.20"
langchain-openai = "^0.3.8"
langchain-text-splitters = "^0.3.6"
langchain-community = "^0.3.19"
pydantic = "^2.10.6"
pypdf = "^5.3.1"
docx2txt = "^0.8"
faiss-cpu = "^1.10.0"
tiktoken = "^0.9.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
